lab2/src/main/java/problem1/Name.java
package problem1;

import java.util.Objects;

public class Name {

  /**
   * Name to store first name, middle name and last name of a person
   */
  private String firstName;
  private String middleName;
  private String lastName;

  /**
   * Constructor, takes 3 params
   * @param firstName String
   * @param middleName String
   * @param lastName String
   */
  public Name(String firstName, String middleName, String lastName) {
    this.firstName = firstName;
    this.middleName = middleName;
    this.lastName = lastName;
  }

  public String getFirstName() {
    return firstName;
  }

  public String getMiddleName() {
    return middleName;
  }

  public String getLastName() {
    return lastName;
  }

  @Override
  public boolean equals(Object o) {
    if (this == o) {return true;}
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Name name = (Name) o;
    return Objects.equals(firstName, name.firstName) && Objects.equals(middleName,
        name.middleName) && Objects.equals(lastName, name.lastName);
  }

  @Override
  public int hashCode() {
    return Objects.hash(firstName, middleName, lastName);
  }

  /** toString method
  @Override
  public String toString() {
    return "problem1.Name{" +
        "firstName='" + firstName + '\'' +
        ", middleName='" + middleName + '\'' +
        ", lastName='" + lastName + '\'' +
        '}';
  }
   */
}
